+++ package.json
@@ -15,7 +15,7 @@
     "@ethersproject/providers": "^5.4.5",
     "@metamask/detect-provider": "^1.2.0",
     "@reduxjs/toolkit": "^1.6.1",
-    "@rigelprotocol_01/default-token-list": "^1.0.0",
+    "@rigelprotocol_01/default-token-list": "^2.0.0",
     "@testing-library/jest-dom": "^5.9.0",
     "@testing-library/react": "^10.2.1",
     "@testing-library/user-event": "^12.0.2",
